Improve homepage functionality and structure for better performance
Refactors the homepage to separate server and client components, moving metadata to a server component and client-side logic to `HomeClient.tsx`. Also updates the hero background image and passes gender/location hints to the API via headers.
